requirements.txt
python3-rpi.gpio
math
RPi.GPIO
time